eccv_github/requirements.txt
absl-py==0.14.1
addict==2.4.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
cleanlab==1.0
click==8.0.1
clip==1.0
cloudpickle==2.1.0
colorama==0.4.4
configparser==5.2.0
contextlib2==21.6.0
cryptography==3.4.7
cycler==0.10.0
DaPy==1.16.1
dassl==0.4.2
dataclasses==0.8
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
docker-pycreds==0.4.0
easydict==1.9
editdistpy==0.1.3
einops==0.4.1
entrypoints==0.3
filelock==3.6.0
flake8==3.7.9
fsspec==2022.2.0
ftfy==6.1.1
future==0.18.2
gdown==3.15.0
gitdb==4.0.9
GitPython==3.1.26
google-auth==1.35.0
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
gql==0.2.0
graphql-core==1.1
grpcio==1.41.0
h5py==3.4.0
hydra-core==1.1.1
hyperopt==0.1.2
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
isort==4.3.21
jedi==0.18.0
Jinja2==3.0.3
joblib==1.0.1
json-tricks==3.15.5
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-packaging==0.11.1
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lightly==1.2.10
lightly-utils==0.0.2
llvmlite==0.33.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
mlconfig==0.1.0
multidict==5.1.0
nbclassic==0.3.5
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nni==2.7
notebook==6.4.8
numba==0.50.1
numpy==1.20.3
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
olefile==0.46
omegaconf==2.1.1
opencv-python==4.5.3.56
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pip==21.0.1
plyfile==0.7.4
pointnet2-ops==3.0.0
prettytable==3.3.0
progressbar==2.5
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyDeprecate==0.3.1
pyflakes==2.1.1
Pygments==2.10.0
PyJWT==2.1.0
pymongo==4.1.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-louvain==0.16
PythonWebHDFS==0.2.3
pytorch-lightning==1.5.10
pytorch-metric-learning==0.9.99
pytorch-model-summary==0.1.2
pytz==2021.3
pyu2f==0.1.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
randaugment==1.0.2
rdflib==6.1.1
referit3d==0.1
regex==2022.1.18
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.20.0
rsa==4.7.2
schema==0.7.5
scikit-image==0.16.2
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.5
setuptools==59.5.0
Shapely==1.8.0
shortuuid==1.0.8
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
subprocess32==3.5.4
symspellpy==6.7.6
tb-nightly==2.9.0a20220405
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.8.30
tomlkit==0.9.2
torch==1.9.1
torch-cluster==1.5.9
torch-geometric==1.7.2
torch-points-kernels==0.6.10
torch-points3d==1.3.0
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-tb-profiler==0.2.1
torchaudio==0.9.0a0+a85b239
torchfile==0.1.0
torchmetrics==0.7.2
torchnet==0.0.4
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
transform3d==0.0.4
transforms3d==0.3.1
typeguard==2.13.3
typing-extensions==3.10.0.2
urllib3==1.26.7
visdom==0.1.8.9
wandb==0.8.36
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==10.3
Werkzeug==2.0.2
wheel==0.37.0
widgetsnbextension==3.5.2
xlrd==2.0.1
xlwt==1.3.0
yacs==0.1.8
yapf==0.29.0
yarl==1.6.3
zipp==3.6.0
